add permissions

+++ app/fs.go
@@ -2,12 +2,13 @@ package app
 
 import (
 	"context"
-	log "github.com/sirupsen/logrus"
-	"golang.org/x/net/webdav"
 	"os"
 	"path"
 	"path/filepath"
 	"strings"
+
+	log "github.com/sirupsen/logrus"
+	"golang.org/x/net/webdav"
 )
 
 // This file is an extension of golang.org/x/net/webdav/file.go.
@@ -58,6 +59,11 @@ func (d Dir) Mkdir(ctx context.Context, name string, perm os.FileMode) error {
 	if name = d.resolve(ctx, name); name == "" {
 		return os.ErrNotExist
 	}
+
+	if !hasWrite(ctx, d) {
+		return os.ErrPermission
+	}
+
 	err := os.Mkdir(name, perm)
 	if err != nil {
 		return err
@@ -78,6 +84,27 @@ func (d Dir) OpenFile(ctx context.Context, name string, flag int, perm os.FileMo
 	if name = d.resolve(ctx, name); name == "" {
 		return nil, os.ErrNotExist
 	}
+
+	if (flag & 0x3) != 0 {
+		if !hasWrite(ctx, d) {
+			return nil, os.ErrPermission
+		}
+	}
+
+	if ((flag & 0x3) == 0) || ((flag & 0x3) == os.O_RDWR) {
+		stat, _ := os.Stat(name)
+
+		if stat != nil && stat.IsDir() {
+			if !hasList(ctx, d) {
+				return nil, os.ErrPermission
+			}
+		}
+
+		if !hasRead(ctx, d) {
+			return nil, os.ErrPermission
+		}
+	}
+
 	f, err := os.OpenFile(name, flag, perm)
 	if err != nil {
 		return nil, err
@@ -103,6 +130,10 @@ func (d Dir) RemoveAll(ctx context.Context, name string) error {
 		return os.ErrInvalid
 	}
 
+	if !hasDelete(ctx, d) {
+		return os.ErrPermission
+	}
+
 	err := os.RemoveAll(name)
 	if err != nil {
 		return err
@@ -126,6 +157,11 @@ func (d Dir) Rename(ctx context.Context, oldName, newName string) error {
 	if newName = d.resolve(ctx, newName); newName == "" {
 		return os.ErrNotExist
 	}
+
+	if !hasWrite(ctx, d) {
+		return os.ErrPermission
+	}
+
 	if root := filepath.Clean(string(d.Config.Dir)); root == oldName || root == newName {
 		// Prohibit renaming from or to the virtual root directory.
 		return os.ErrInvalid
@@ -152,5 +188,46 @@ func (d Dir) Stat(ctx context.Context, name string) (os.FileInfo, error) {
 	if name = d.resolve(ctx, name); name == "" {
 		return nil, os.ErrNotExist
 	}
+
+	if !hasList(ctx, d) {
+		return nil, os.ErrPermission
+	}
+
 	return os.Stat(name)
+}
+
+func hasRead(ctx context.Context, d Dir) bool {
+	authInfo := AuthFromContext(ctx)
+	if authInfo != nil && authInfo.Authenticated {
+		userInfo := d.Config.Users[authInfo.Username]
+		return userInfo.Read == nil || *userInfo.Read
+	}
+	return len(d.Config.Users) == 0 /* no users, no permission to check */
+}
+
+func hasWrite(ctx context.Context, d Dir) bool {
+	authInfo := AuthFromContext(ctx)
+	if authInfo != nil && authInfo.Authenticated {
+		userInfo := d.Config.Users[authInfo.Username]
+		return userInfo.Write == nil || *userInfo.Write
+	}
+	return len(d.Config.Users) == 0 /* no users, no permission to check */
+}
+
+func hasDelete(ctx context.Context, d Dir) bool {
+	authInfo := AuthFromContext(ctx)
+	if authInfo != nil && authInfo.Authenticated {
+		userInfo := d.Config.Users[authInfo.Username]
+		return userInfo.Delete == nil || *userInfo.Delete
+	}
+	return len(d.Config.Users) == 0 /* no users, no permission to check */
+}
+
+func hasList(ctx context.Context, d Dir) bool {
+	authInfo := AuthFromContext(ctx)
+	if authInfo != nil && authInfo.Authenticated {
+		userInfo := d.Config.Users[authInfo.Username]
+		return userInfo.List == nil || *userInfo.List
+	}
+	return len(d.Config.Users) == 0 /* no users, no permission to check */
 }
